basic autorisasi

+++ app/Http/Controllers/PostController.php
@@ -70,7 +70,7 @@ class PostController extends Controller
         $attr['slug'] = \Str::slug(request('title'));
         $attr['category_id'] = request('category');
         // create new post
-        $post = Post::create($attr);
+        $post = auth()->user()->posts()->create($attr);
 
         $post->tags()->attach(request('tags'));
 
@@ -104,10 +104,15 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        $post->tags()->detach();
-        $post->delete();
-        session()->flash('success', 'The post was destroyed');
-        return redirect('post');
+        if (auth()->user()->is($post->author)) {
+            $post->tags()->detach();
+            $post->delete();
+            session()->flash('success', 'The post was destroyed');
+            return redirect('post');
+        } else {
+            session()->flash("error", "It wasn't your post");
+            return redirect('post');
+        }
     }
 
     // public function validateRequest()
